server/backend/requirements.txt
# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 이미지 처리
Pillow==10.1.0
numpy==1.24.3
tifffile

# 데이터 저장 (precomputed_writer.py가 사용)
zarr
Form
tifffile

# 시스템 모니터링
psutil==5.9.8

# HTTP 클라이언트 (헬스체크용)
requests==2.31.0
httpx==0.25.2

# 테스트 도구
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

